+++ src/app/_components/review/ReviewItem/ReviewItemHeader.tsx
@@ -5,8 +5,10 @@ import { convertTime } from '@/utils/convertTime';
 
 type ReviewItemHeaderProps = Pick<
   ReviewItemType,
-  'profileImageUrl' | 'nickName' | 'rating' | 'createdAt'
->;
+  'profileImageUrl' | 'nickName' | 'createdAt'
+> & {
+  rating?: number;
+};
 
 export default function ReviewItemHeader({
   profileImageUrl,
@@ -18,11 +20,11 @@ export default function ReviewItemHeader({
     <div className="flex items-center justify-between px-8">
       <div className="flex items-center gap-2">
         <UserImage src={profileImageUrl} />
-        <span className="text-14 font-bold">{nickName}</span>
+        <span className="text-16 font-bold">{nickName}</span>
       </div>
       <div className="flex flex-col">
         <div className="flex gap-1 items-center justify-end">
-          <Icon name="star" size={8} />
+          {rating && <Icon name="star" size={8} />}
           <span className="text-12">{rating}</span>
         </div>
         <span className="text-12 text-text_grey">{convertTime(createdAt)}</span>
